Generalize withoutBalanceChangeIt
